test: NewWorksForYou
